Refactor markdown typography styles and code block handling
- Customize `.prose` in `src/styles/global.css` to use brand colors (Teal), serif headings (Merriweather), and improved spacing for an editorial look.
- Implement a dark theme for code blocks with rounded corners.
- Extract copy-to-clipboard functionality to `src/scripts/code-copy.ts` and apply it globally via `src/scripts/layout.ts`, ensuring it works on Blog, Apps, and Devlog pages.
- Add unit tests for the code copy logic in `src/scripts/code-copy.test.ts`.
- Remove inline styles from `src/pages/blog/[...slug].astro` in favor of global styles.
