+++ src/pages/blog/[...slug].astro
@@ -252,36 +252,4 @@ const breadcrumbItems = [
   </article>
 </Layout>
 
-<style is:global>
-  /* Copy Button Styles */
-  .copy-code-btn {
-    position: absolute;
-    top: 0.5rem;
-    right: 0.5rem;
-    padding: 0.25rem;
-    background-color: rgba(255, 255, 255, 0.1);
-    border: 1px solid rgba(255, 255, 255, 0.2);
-    border-radius: 0.375rem;
-    color: #e5e7eb;
-    cursor: pointer;
-    transition: all 0.2s;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .copy-code-btn:hover {
-    background-color: rgba(255, 255, 255, 0.2);
-    color: white;
-  }
-
-  .copy-code-btn .material-icons {
-    font-size: 1.25rem;
-  }
-
-  .prose pre {
-    position: relative;
-  }
-</style>
-
 <script src="../../scripts/blog.ts"></script>
